Project Init 2

+++ DemoFreeImage/DemoFreeImage/main.cpp
@@ -1,21 +1,23 @@
 #include <iostream>
 #include "FreeImage.h"
+#include "Vector3.h"
 
 int main(int argc, char** argv)
 {
+
 	RGBQUAD color;
 	FIBITMAP* image;
 	image = FreeImage_Allocate(256, 256, 32);
 
-
 	color.rgbGreen = 0;
+
 	for (size_t i = 0; i < 640; i++)
 	{
-		color.rgbRed = i;
+		color.rgbRed = static_cast<BYTE>(i);
 
 		for (size_t j = 0; j < 480; j++)
 		{
-			color.rgbBlue = j;
+			color.rgbBlue = static_cast<BYTE>(j);
 
 			FreeImage_SetPixelColor(image, i, j, &color);
 		}
@@ -24,4 +26,5 @@ int main(int argc, char** argv)
 	FreeImage_Save(FIF_BMP, image, "out.bmp");
 
 	return EXIT_SUCCESS;
+
 }
